Working on R2000: use CADBitStreamReader

+++ lib/cadfile.h
@@ -68,7 +68,7 @@ public:
     const CADTables & getTables() const;
 
 public:
-    virtual int    ParseFile( enum OpenOptions eOptions, bool bReadUnsupportedGeometries = true );
+    virtual int32_t    ParseFile( enum OpenOptions eOptions, bool bReadUnsupportedGeometries = true );
     virtual size_t GetLayersCount() const;
     virtual CADLayer& GetLayer( size_t index );
 
@@ -88,7 +88,7 @@ protected:
      * @param bHandlesOnly set TRUE if object data should be skipped, and only object handles should be read.
      * @return pointer to CADObject or nullptr. User have to free returned pointer.
      */
-    virtual CADObject * GetObject( long dObjectHandle, bool bHandlesOnly = false ) = 0;
+    virtual CADObject * GetObject( int32_t dObjectHandle, bool bHandlesOnly = false ) = 0;
 
     /**
      * @brief read geometry from CAD file
@@ -97,40 +97,40 @@ protected:
      * @param handle Handle of BlockRef (0 if geometry is not in block reference)
      * @return NULL if failed or pointer which mast be feed by user
      */
-    virtual CADGeometry * GetGeometry( size_t iLayerIndex, long dHandle, long dBlockRefHandle = 0 ) = 0;
+    virtual CADGeometry * GetGeometry( size_t iLayerIndex, int32_t dHandle, int32_t dBlockRefHandle = 0 ) = 0;
 
     /**
      * @brief initially read some basic values and section locator
      * @return CADErrorCodes::SUCCESS if OK, or error code
      */
-    virtual int ReadSectionLocators() = 0;
+    virtual int32_t ReadSectionLocators() = 0;
 
     /**
      * @brief Read header from CAD file
      * @param eOptions Read options
      * @return CADErrorCodes::SUCCESS if OK, or error code
      */
-    virtual int ReadHeader( enum OpenOptions eOptions ) = 0;
+    virtual int32_t ReadHeader( enum OpenOptions eOptions ) = 0;
 
     /**
      * @brief Read classes from CAD file
      * @param eOptions Read options
      * @return CADErrorCodes::SUCCESS if OK, or error code
      */
-    virtual int ReadClasses( enum OpenOptions eOptions ) = 0;
+    virtual int32_t ReadClasses( enum OpenOptions eOptions ) = 0;
 
     /**
      * @brief Create the file map for fast access to CAD objects
      * @return CADErrorCodes::SUCCESS if OK, or error code
      */
-    virtual int CreateFileMap() = 0;
+    virtual int32_t CreateFileMap() = 0;
 
     /**
      * @brief Read tables from CAD file
      * @param eOptions Read options
      * @return CADErrorCodes::SUCCESS if OK, or error code
      */
-    virtual int ReadTables( enum OpenOptions eOptions );
+    virtual int32_t ReadTables( enum OpenOptions eOptions );
 
     /**
      * @brief returns value of flag Read Unsupported Geometries
@@ -144,7 +144,7 @@ protected:
     CADTables  oTables;
 
 protected:
-    std::map<long, long> mapObjects; // object index <-> file offset
+    std::map<int32_t, int32_t> mapObjects; // object index <-> file offset
     bool bReadingUnsupportedGeometries;
 };
